// frontend/src/client/types.gen.ts
export type ActorCreate = {
    /**
     * First Name
     */
    first_name: string;
    /**
     * Last Name
     */
    last_name: string;
    /**
     * Birth Date
     */
    birth_date?: string | null;
};
// ...
export type DirectorCreate = {
    /**
     * First Name
     */
    first_name: string;
    /**
     * Last Name
     */
    last_name: string;
    /**
     * Birth Date
     */
    birth_date?: string | null;
};
// ...
export type MovieCreate = {
    /**
     * Title
     */
    title: string;
    /**
     * Synopsis
     */
    synopsis: string;
    /**
     * Release Year
     */
    release_year: number;
    /**
     * Duration
     */
    duration: number;
    /**
     * Age Rating
     */
    age_rating: 'G' | 'PG' | 'PG-13' | 'R' | 'NC-17';
    /**
     * Poster Url
     */
    poster_url: string;
    /**
     * Trailer Url
     */
    trailer_url: string;
    /**
     * Genre Ids
     */
    genre_ids: Array<string>;
    /**
     * Actors
     */
    actors: Array<ActorCreate>;
    /**
     * Directors
     */
    directors: Array<DirectorCreate>;
};